move components for portfolios page into components folder
