+++ apps/ui/stratify-ui/app/(screens)/app/markets/AssetSearch/AssetSearch.test.tsx
@@ -27,6 +27,7 @@ const mockIsNoResultsFound = vi.fn();
 const mockUseAssetSearch = () => ({
     searchResults: [
         {
+            id: 1,
             name: "Apple Inc.",
             symbol: "AAPL",
             assetType: "STOCK",
