+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/AddInvestment/AddInvestmentModal.test.tsx
