+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/AddInvestment/add-investment-schema.test.ts
