+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/AddTrade/AddTradeModal.tsx
@@ -12,44 +12,38 @@ import {
 import { useQueryClient } from "@tanstack/react-query";
 import { X } from "lucide-react";
 import { useEffect, useState } from "react";
-import {
-    AddInvestmentSchema,
-    addInvestmentSchema,
-} from "./add-investment-schema";
-import { Field, FieldLabel } from "@/app/components/ui/field";
-import { useDebouncedValue } from "@tanstack/react-pacer";
-import {
-    SearchAsset,
-    useAssetSearch,
-} from "../../markets/AssetSearch/useAssetSearch";
-import {
-    Command,
-    CommandInput,
-    CommandList,
-} from "@/app/components/ui/command";
-import { Skeleton } from "@/app/components/ui/skeleton";
-import { cn } from "@/lib/utils";
-import AssetNameCard from "./AssetNameCard";
-import { useSessionContext } from "../../SessionProvider";
-import { useHistoricAssetPrice } from "../hooks/useHistoricAssetPrice";
-import { useHistoricCurrencyPairPrice } from "../hooks/useHistoricConversionRate";
+import { useSessionContext } from "../../../SessionProvider";
+import { useHistoricAssetPrice } from "../../hooks/useHistoricAssetPrice";
+import { useHistoricCurrencyPairPrice } from "../../hooks/useHistoricConversionRate";
 import {
     AddTradeErrorResponse,
     AddTradeRequestSchema,
     useAddTrade,
-} from "../hooks/useAddTrade";
+} from "../../hooks/useAddTrade";
 import { HTTPError } from "ky";
 import { toast } from "sonner";
+import { addTradeSchema, AddTradeSchema } from "./add-trade-schema";
+import { Field, FieldLabel } from "@/app/components/ui/field";
+import { Tabs, TabsList, TabsTrigger } from "@/app/components/ui/tabs";
 
-export interface AddInvestmentModalProps {
+interface InvestmentDetails {
+    assetId: number;
+    symbol: string;
+    name: string;
+    assetCurrency: string | null;
+}
+
+export interface AddTradeModalProps {
     portfolioId: number;
+    investment: InvestmentDetails;
     isOpen: boolean;
     handleClose: () => void;
 }
 
-const defaultValues: AddInvestmentSchema = {
+const defaultValues: AddTradeSchema = {
     assetName: "",
     quantity: "",
+    tradeType: "BUY",
     tradeDate: "",
     pricePerShare: "",
     currencyConversionRate: "1",
@@ -58,59 +52,39 @@ const defaultValues: AddInvestmentSchema = {
     assetCurrencyTotal: 0,
 };
 
-const addInvestmentFormOptions = formOptions({
-    formId: "add-investment-form",
+const addTradeFormOptions = formOptions({
+    formId: "add-trade-form",
     defaultValues,
 });
 
-const AddInvestmentModal = ({
+const AddTradeModal = ({
     portfolioId,
+    investment,
     isOpen,
     handleClose,
-}: AddInvestmentModalProps) => {
+}: AddTradeModalProps) => {
+    const { assetId, name, symbol, assetCurrency } = investment;
     const [isSubmitDisabled, setIsSubmitDisabled] = useState(true);
     const queryClient = useQueryClient();
 
     const { session } = useSessionContext();
     const userCurrency = session?.userDetails.currency ?? "---";
 
-    const [searchValue, setSearchValue] = useState("");
-    const [selectedAsset, setSelectedAsset] = useState<SearchAsset | null>(
-        null,
-    );
-    const [isAssetListOpen, setIsAssetListOpen] = useState(false);
-
-    const debouncedSearchValue = useDebouncedValue(searchValue, {
-        wait: 500,
-    })[0];
-
-    const {
-        searchResults,
-        isSearching,
-        search,
-        isNoResultsFound,
-        searchStatus,
-        resetSearch,
-    } = useAssetSearch(debouncedSearchValue);
-
     const { mutate: addTrade, isPending: isAddingTrade } =
         useAddTrade(portfolioId);
 
-    useEffect(() => {
-        if (debouncedSearchValue.length > 0) {
-            search();
-        }
-    }, [debouncedSearchValue, search]);
+    const [isAttemptingToSellMoreThanHeld, setIsAttemptingToSellMoreThanHeld] =
+        useState(false);
 
     const form = useAppForm({
-        ...addInvestmentFormOptions,
+        ...addTradeFormOptions,
         validators: {
-            onChange: addInvestmentSchema,
+            onChange: addTradeSchema,
             onBlurAsync: async ({ value }) => {
-                const result = addInvestmentSchema.safeParse(value);
+                const result = addTradeSchema.safeParse(value);
 
                 //? If currency conversion is required but the conversion rate isn't set, then the submit button should be disabled
-                if (selectedAsset?.currency !== userCurrency) {
+                if (assetCurrency !== userCurrency) {
                     const currencyConversionRateValue =
                         value.currencyConversionRate
                             ? parseFloat(value.currencyConversionRate)
@@ -130,17 +104,19 @@ const AddInvestmentModal = ({
             },
         },
         onSubmit: async ({ value }) => {
-            if (!selectedAsset) return;
+            if (!value.tradeType) {
+                return toast.error("Please select a trade type.");
+            }
 
             const requestBody = {
-                assetId: selectedAsset.id,
+                assetId,
                 quantity: parseFloat(value.quantity),
                 pricePerShare: parseFloat(value.pricePerShare),
                 currencyConversionRate: parseFloat(
                     value.currencyConversionRate,
                 ),
                 fee: parseFloat(value.fee || "0"),
-                tradeAction: "BUY",
+                tradeAction: value.tradeType,
                 totalAmount: value.total,
                 assetCurrencyTotalAmount: value.assetCurrencyTotal || 0,
                 tradeDate: value.tradeDate,
@@ -162,10 +138,11 @@ const AddInvestmentModal = ({
                     });
 
                     toast.success(
-                        "Investment added to your portfolio successfully.",
+                        "Trade added to your portfolio successfully.",
                     );
 
-                    closeModal();
+                    handleClose();
+                    form.reset();
                 },
                 onError: async (error) => {
                     const httpError = error as HTTPError;
@@ -182,13 +159,18 @@ const AddInvestmentModal = ({
                             );
                             setIsSubmitDisabled(true);
                         }
+
+                        if (errorMessage === "cannotSellMoreThanHeld") {
+                            setIsAttemptingToSellMoreThanHeld(true);
+                            setIsSubmitDisabled(true);
+                        }
                     }
                 },
             });
         },
         onSubmitInvalid: () => {
             toast.error(
-                "Investment could not be added to your portfolio. Please check the form for errors.",
+                "Trade could not be added to your portfolio. Please check the form for errors.",
             );
             setIsSubmitDisabled(true);
         },
@@ -200,14 +182,12 @@ const AddInvestmentModal = ({
         data: historicAssetPrice,
         mutate: fetchHistoricAssetPrice,
         isPending: isFetchingHistoricPrice,
-        reset: resetHistoricAssetPrice,
     } = useHistoricAssetPrice();
 
     const {
         data: historicCurrencyPairPrice,
         mutate: fetchHistoricCurrencyPairPrice,
         isPending: isFetchingHistoricCurrencyPairPrice,
-        reset: resetHistoricCurrencyPairPrice,
     } = useHistoricCurrencyPairPrice();
 
     useEffect(() => {
@@ -218,13 +198,13 @@ const AddInvestmentModal = ({
             );
         }
 
-        if (selectedAsset?.currency === "GBX" && userCurrency === "GBP") {
+        if (assetCurrency === "GBX" && userCurrency === "GBP") {
             form.setFieldValue("currencyConversionRate", "0.01");
         }
 
         if (historicCurrencyPairPrice) {
             const conversionRate =
-                selectedAsset?.currency === "GBX"
+                assetCurrency === "GBX"
                     ? parseFloat(historicCurrencyPairPrice.price) / 100
                     : parseFloat(historicCurrencyPairPrice.price);
 
@@ -233,26 +213,21 @@ const AddInvestmentModal = ({
                 conversionRate.toString(),
             );
         }
-    }, [
-        historicAssetPrice,
-        historicCurrencyPairPrice,
-        form,
-        selectedAsset?.currency,
-    ]);
+    }, [historicAssetPrice, historicCurrencyPairPrice, form, assetCurrency]);
 
     const pricePerShare = formValues.pricePerShare
         ? parseFloat(formValues.pricePerShare)
         : 0;
     const quantity = formValues.quantity ? parseFloat(formValues.quantity) : 0;
+
+    const tradeType = formValues.tradeType;
     const currencyConversionRate = formValues.currencyConversionRate
         ? parseFloat(formValues.currencyConversionRate)
         : 1;
 
     const fee = formValues.fee ? parseFloat(formValues.fee) : 0;
-    const assetCurrency = selectedAsset?.currency ?? "---";
 
-    const isCurrencyConversionRequired =
-        selectedAsset && assetCurrency !== userCurrency;
+    const isCurrencyConversionRequired = assetCurrency !== userCurrency;
 
     const assetCurrencySubtotal =
         pricePerShare > 0 && quantity > 0 ? pricePerShare * quantity : 0;
@@ -271,18 +246,14 @@ const AddInvestmentModal = ({
         }
     }, [total, assetCurrencySubtotal, isCurrencyConversionRequired]);
 
-    const closeModal = () => {
-        setSearchValue("");
-        setSelectedAsset(null);
-        setIsSubmitDisabled(true);
-        form.reset();
-        resetHistoricAssetPrice();
-        resetHistoricCurrencyPairPrice();
-        handleClose();
-    };
-
     return (
-        <Dialog open={isOpen} onOpenChange={() => closeModal()}>
+        <Dialog
+            open={isOpen}
+            onOpenChange={() => {
+                form.reset();
+                handleClose();
+            }}
+        >
             <DialogContent className="bg-muted-lightest border border-primary-dark font-sans">
                 <DialogHeader>
                     <div className="flex flex-row items-center justify-between">
@@ -290,12 +261,15 @@ const AddInvestmentModal = ({
                             className="font-medium text-lg leading-7 text-secondary-darker"
                             data-testid="modal-title"
                         >
-                            Add Investment
+                            Add Trade
                         </DialogTitle>
                         <X
                             size={20}
                             className="cursor-pointer text-muted-dark hover:text-primary-darker transition-colors"
-                            onClick={() => closeModal()}
+                            onClick={() => {
+                                handleClose();
+                                form.reset();
+                            }}
                             data-testid="close-modal-icon"
                         />
                     </div>
@@ -303,9 +277,8 @@ const AddInvestmentModal = ({
                         className="text-secondary-darker leading-5"
                         data-testid="modal-description"
                     >
-                        Add a new investment to monitor its performance and see
-                        how it impacts your portfolio alongside other
-                        investments.
+                        Add a transaction to keep your portfolio up to date and
+                        monitor it with more accuracy.
                     </DialogDescription>
                 </DialogHeader>
                 <form
@@ -316,137 +289,30 @@ const AddInvestmentModal = ({
                     className="flex flex-col gap-y-2.5"
                 >
                     <form.AppField name="assetName">
-                        {(field) => {
-                            return (
-                                <Field
-                                    className="flex flex-col gap-y-1.5"
-                                    data-testid="asset-name-field"
-                                >
-                                    <FieldLabel htmlFor="assetName">
-                                        <span className="flex flex-row w-full items-center justify-between">
-                                            Asset Name
-                                            {field.state.value && (
-                                                <span
-                                                    className="p-1 rounded-lg bg-secondary-lighter font-sans font-semibold text-secondary-darker cursor-pointer hover:bg-secondary-base hover:text-secondary-lightest transition-colors"
-                                                    onClick={() => {
-                                                        setSearchValue("");
-                                                        setSelectedAsset(null);
-                                                        setIsAssetListOpen(
-                                                            false,
-                                                        );
-                                                        setIsSubmitDisabled(
-                                                            true,
-                                                        );
-                                                        resetHistoricAssetPrice();
-                                                        resetHistoricCurrencyPairPrice();
-                                                        form.reset();
-                                                    }}
-                                                >
-                                                    Clear
-                                                </span>
-                                            )}
-                                        </span>
-                                    </FieldLabel>
-                                    <Command>
-                                        <CommandInput
-                                            id="assetName"
-                                            placeholder="Search for an asset by name or symbol"
-                                            className="bg-white border border-secondary-dark rounded-md focus-visible:ring-secondary-dark"
-                                            inputClassName="placeholder:text-secondary-light text-secondary-dark"
-                                            showIcon={false}
-                                            iconClassName="text-secondary-dark"
-                                            disabled={field.state.value !== ""}
-                                            value={
-                                                field.state.value
-                                                    ? `${selectedAsset?.name} (${selectedAsset?.symbol})`
-                                                    : searchValue
-                                            }
-                                            onValueChange={(searchValue) =>
-                                                setSearchValue(searchValue)
-                                            }
-                                            onFocus={() =>
-                                                setIsAssetListOpen(true)
-                                            }
-                                        />
-                                        {searchStatus != "idle" && (
-                                            <CommandList
-                                                className={cn(
-                                                    "p-1 gap-y-1.5 bg-white border border-secondary-dark rounded-md mt-1",
-                                                    isAssetListOpen
-                                                        ? ""
-                                                        : "hidden",
-                                                )}
-                                                asChild
-                                            >
-                                                {isSearching ? (
-                                                    <div className="flex flex-col gap-2">
-                                                        {Array.from({
-                                                            length: 3,
-                                                        }).map((_, index) => (
-                                                            <Skeleton
-                                                                key={index}
-                                                                className="w-full h-10 rounded-xl bg-secondary-lighter"
-                                                            />
-                                                        ))}
-                                                    </div>
-                                                ) : (
-                                                    <div className="flex flex-col gap-y-1">
-                                                        {searchResults?.map(
-                                                            (asset) => (
-                                                                <AssetNameCard
-                                                                    key={`${asset.symbol}-${asset.name}`}
-                                                                    asset={
-                                                                        asset
-                                                                    }
-                                                                    onSelect={() => {
-                                                                        field.handleChange(
-                                                                            asset.name,
-                                                                        );
-                                                                        setSelectedAsset(
-                                                                            asset,
-                                                                        );
-                                                                        setIsAssetListOpen(
-                                                                            false,
-                                                                        );
-
-                                                                        resetSearch();
-                                                                    }}
-                                                                />
-                                                            ),
-                                                        )}
-                                                    </div>
-                                                )}
-                                                {isNoResultsFound &&
-                                                    !searchResults &&
-                                                    searchStatus ===
-                                                        "success" && (
-                                                        <div className="p-2 text-center text-base font-sans text-secondary-dark">
-                                                            No assets found.
-                                                        </div>
-                                                    )}
-                                            </CommandList>
-                                        )}
-                                    </Command>
-                                </Field>
-                            );
-                        }}
+                        {({ TextInput }) => (
+                            <TextInput
+                                id="assetName"
+                                label="Asset Name"
+                                dataTestId="asset-name-field"
+                                defaultValue={`${name} (${symbol})`}
+                                disabled
+                            />
+                        )}
                     </form.AppField>
                     <form.AppField
                         name="tradeDate"
                         listeners={{
                             onChange: ({ value }) => {
-                                if (selectedAsset) {
-                                    fetchHistoricAssetPrice({
-                                        assetId: selectedAsset.id,
+                                fetchHistoricAssetPrice({
+                                    assetId,
+                                    tradeDate: value,
+                                });
+
+                                if (isCurrencyConversionRequired) {
+                                    fetchHistoricCurrencyPairPrice({
+                                        currencyPair: `${assetCurrency === "GBX" ? "GBP" : assetCurrency}${userCurrency}`,
                                         tradeDate: value,
                                     });
-
-                                    if (isCurrencyConversionRequired) {
-                                        fetchHistoricCurrencyPairPrice({
-                                            currencyPair: `${selectedAsset.currency === "GBX" ? "GBP" : selectedAsset.currency}${userCurrency}`,
-                                            tradeDate: value,
-                                        });
-                                    }
                                 }
                             },
                         }}
@@ -476,9 +342,7 @@ const AddInvestmentModal = ({
                                     dataTestId="price-per-share-field"
                                     label="Price per Share"
                                     isLoading={isFetchingHistoricPrice}
-                                    currencyCode={
-                                        selectedAsset?.currency ?? "---"
-                                    }
+                                    currencyCode={assetCurrency ?? "---"}
                                     placeholder="Enter price per share"
                                     inputClassName=""
                                     error={
@@ -490,24 +354,71 @@ const AddInvestmentModal = ({
                             );
                         }}
                     </form.AppField>
-                    <form.AppField name="quantity">
-                        {({ state: { meta }, TextInput }) => {
-                            return (
-                                <TextInput
-                                    id="quantity"
-                                    dataTestId="quantity-field"
-                                    label="Number of Shares"
-                                    placeholder="Enter number of shares"
-                                    inputClassName="bg-white text-secondary-dark"
-                                    error={
-                                        meta.isTouched
-                                            ? meta.errors?.[0]?.message
-                                            : undefined
-                                    }
-                                />
-                            );
-                        }}
-                    </form.AppField>
+                    <div className="flex flex-row gap-x-4">
+                        <form.AppField
+                            name="quantity"
+                            validators={{
+                                onChange: () =>
+                                    setIsAttemptingToSellMoreThanHeld(false),
+                            }}
+                        >
+                            {({ state: { meta }, TextInput }) => {
+                                const validationError = meta.isTouched
+                                    ? meta.errors?.[0]?.message
+                                    : undefined;
+
+                                const quantityError =
+                                    isAttemptingToSellMoreThanHeld
+                                        ? "You cannot sell more shares than what you currently hold. Please adjust the quantity or change the trade type to Buy."
+                                        : validationError;
+                                return (
+                                    <TextInput
+                                        id="quantity"
+                                        dataTestId="quantity-field"
+                                        label="Number of Shares"
+                                        placeholder="Enter number of shares"
+                                        inputClassName="bg-white text-secondary-dark"
+                                        error={quantityError}
+                                    />
+                                );
+                            }}
+                        </form.AppField>
+                        <form.AppField name="tradeType">
+                            {(field) => (
+                                <Field
+                                    className="flex flex-col gap-y-1.5"
+                                    data-testid="trade-type-field"
+                                >
+                                    <FieldLabel htmlFor="tradeType">
+                                        Trade Type
+                                    </FieldLabel>
+                                    <Tabs
+                                        defaultValue="BUY"
+                                        onValueChange={(value) =>
+                                            field.setValue(
+                                                value as "BUY" | "SELL",
+                                            )
+                                        }
+                                    >
+                                        <TabsList className="bg-white border border-secondary-dark rounded-md w-full">
+                                            <TabsTrigger
+                                                value="BUY"
+                                                className="rounded-md p-1 text-sm font-normal text-muted-dark hover:bg-positive-lighter hover:text-positive-dark hover:font-medium data-[state=active]:bg-positive-light data-[state=active]:text-positive-darker data-[state=active]:font-semibold"
+                                            >
+                                                Buy
+                                            </TabsTrigger>
+                                            <TabsTrigger
+                                                value="SELL"
+                                                className="rounded-md p-1 text-sm font-normal text-muted-dark hover:bg-negative-lighter hover:text-negative-dark hover:font-medium data-[state=active]:bg-negative-light data-[state=active]:text-negative-darker data-[state=active]:font-semibold"
+                                            >
+                                                Sell
+                                            </TabsTrigger>
+                                        </TabsList>
+                                    </Tabs>
+                                </Field>
+                            )}
+                        </form.AppField>
+                    </div>
                     {isCurrencyConversionRequired ? (
                         <form.AppField name="currencyConversionRate">
                             {({ state: { meta }, CurrencyInput }) => {
@@ -518,7 +429,7 @@ const AddInvestmentModal = ({
                                         isLoading={
                                             isFetchingHistoricCurrencyPairPrice
                                         }
-                                        currencyCode={`${selectedAsset?.currency}/${userCurrency}`}
+                                        currencyCode={`${assetCurrency ?? "---"}/${userCurrency}`}
                                         placeholder="Enter currency conversion rate"
                                         error={
                                             meta.isTouched
@@ -588,7 +499,8 @@ const AddInvestmentModal = ({
                         </div>
                         <div className="flex flex-row items-center justify-between mt-2">
                             <span className="text-secondary-dark font-semibold">
-                                Total
+                                Total {tradeType === "BUY" && "Bought"}
+                                {tradeType === "SELL" && "Sold"}
                             </span>
                             <span className="font-medium text-muted-darker">
                                 {total > 0 && userCurrency
@@ -604,7 +516,7 @@ const AddInvestmentModal = ({
                                 ]}
                             >
                                 <form.SubmitButton
-                                    label="Add Investment"
+                                    label={`Add Trade for ${symbol}`}
                                     className="w-full"
                                     isDisabled={
                                         !form.state.canSubmit ||
@@ -623,4 +535,4 @@ const AddInvestmentModal = ({
     );
 };
 
-export default AddInvestmentModal;
+export default AddTradeModal;
