+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/AddTrade/add-trade-schema.ts
