+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/AssetAllocationCard/placeholderChartData.ts
