+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/InvestmentsTable/useInvestmentsList.ts
