+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/PortfolioMetrics/_mocks/mockMetricsData.ts
@@ -1,0 +1,14 @@
+import { PortfolioMetricsResponse } from "../usePortfolioMetrics";
+
+export const mockMetricsData = {
+    totalValue: 100000,
+    overallReturn: {
+        absolute: 5000,
+        percentage: 5,
+    },
+    riskMetrics: {
+        volatility: 10,
+        sortinoRatio: 1.5,
+        riskLevel: "low",
+    },
+} satisfies PortfolioMetricsResponse;
