+++ apps/ui/stratify-ui/app/(screens)/app/portfolios/components/PortfolioValueChart/PortfolioValueDetails.tsx
